apps/worker/action/requirements.txt
fastapi>=0.115.0
uvicorn>=0.30.0
pydantic>=2.9.0
python-json-logger>=2.0.7
google-api-python-client>=2.118.0
google-auth>=2.28.1
google-generativeai>=0.8.0
httpx>=0.27.0
python-dotenv>=1.0.0
